% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pivot_table.R
\name{rRound}
\alias{rRound}
\title{Title Rounds the count number}
\usage{
rRound(cnt, seed_n)
}
\arguments{
\item{cnt}{input number.}

\item{seed_n}{seed.}
}
\value{
round number
}
\description{
Title Rounds the count number
}
