% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pivot_table.R
\name{shinypivottabler2}
\alias{shinypivottabler2}
\alias{shinypivottablerUI2}
\title{Shiny module to render and export pivot tables.}
\usage{
shinypivottabler2(
  input,
  output,
  session,
  data,
  pivot_cols = NULL,
  indicator_cols = NULL,
  remove_indicator = NULL,
  max_n_pivot_cols = 100,
  additional_expr_num = list(),
  additional_expr_char = list(),
  additional_combine = list(),
  theme = NULL,
  export_styles = TRUE,
  show_title = TRUE,
  initialization = NULL
)

shinypivottablerUI2(
  id,
  app_colors = c("#59bb28", "#217346"),
  app_linewidth = 8
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny input}

\item{data}{\code{data.frame} / \code{data.table}. Initial data table.}

\item{pivot_cols}{\code{character} (NULL). Columns to be used as pivot in rows and cols.}

\item{indicator_cols}{\code{character} (NULL). Columns on which indicators will be calculated.}

\item{remove_indicator}{\code{list} (NULL). The indicators do no use.}

\item{max_n_pivot_cols}{\code{numeric} (100). Maximum unique values for a \code{pivot_cols} if pivot_cols = NULL}

\item{additional_expr_num}{\code{named list} (list()). Additional computations to be allowed for quantitative vars.}

\item{additional_expr_char}{\code{named list} (list()). Additional computations to be allowed for qualitative vars.}

\item{additional_combine}{\code{named list} (list()). Additional combinations to be allowed.}

\item{theme}{\code{list} (NULL). Theme to customize the output of the pivot table.}

\item{export_styles}{\code{boolean} (TRUE). Whether or not to apply styles (like the theme) when exporting to Excel.}

\item{show_title}{\code{boolean} (TRUE). Whether or not to display the app title.
Some styles may not be supported by Excel.}

\item{initialization}{\code{named list} (NULL). Initialization parameters to display a table when launching the module.
Available fields are :
\itemize{
\item{\code{rows:}}{ Selected pivot rows.}
\item{\code{cols:}}{ Selected pivot columns.}
\item{\code{target, combine target:}} { Selected target and combine_target columns.}.
\item{\code{idc, combine_idc:}}{ Selected idc and combine_idc columns.}
\item{\code{combine:}}{ Selected combine operator.}
\item{\code{format_digit, format_prefix, format_suffix, format_sep_thousands, format_decimal:}}{ Selected formats for the table idc.}
\item{\code{idcs:}}{ idcs to be displayed (list of named list), see the example to get the fields.}
}}

\item{id}{\code{character}. An ID string}

\item{app_colors}{\code{character}. Vector of two colors c("#59bb28", "#217346") (borders)}

\item{app_linewidth}{\code{numeric}. Borders width}
}
\value{
Nothing. Just Start a Shiny module.
}
\description{
Shiny module to render and export pivot tables.
}
